requirements.txt
Django~=2.2.4
djangorestframework
python-interface
paramiko
celery
redis
pandas
python-dateutil
